ColdBore/data_qc_tracker_project/requirements.txt
streamlit==1.22.0
pandas==1.5.3
plotly==5.14.1
psycopg2-binary==2.9.6
openpyxl==3.1.2
